pyproject.toml
[project]
name = "ftrader"
version = "0.1.0"
description = "币安马丁格尔交易框架"
requires-python = ">=3.9"
dependencies = [
    "ccxt>=4.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

